requirements.txt
yfinance
pandas
ta
pyyaml
